Sync bnd project

+++ bndtools.builder/src/org/bndtools/builder/MarkerSupport.java
@@ -3,8 +3,10 @@ package org.bndtools.builder;
 import static org.bndtools.api.BndtoolsConstants.BNDTOOLS_MARKER_CONTEXT_ATTR;
 import static org.bndtools.api.BndtoolsConstants.BNDTOOLS_MARKER_FILE_ATTR;
 import static org.bndtools.api.BndtoolsConstants.BNDTOOLS_MARKER_HEADER_ATTR;
+import static org.bndtools.api.BndtoolsConstants.BNDTOOLS_MARKER_PROJECT_ATTR;
 import static org.bndtools.api.BndtoolsConstants.BNDTOOLS_MARKER_REFERENCE_ATTR;
 import static org.bndtools.api.BndtoolsConstants.CORE_PLUGIN_ID;
+import static org.bndtools.api.BndtoolsConstants.MARKER_BND_BLOCKER;
 import static org.bndtools.api.BndtoolsConstants.MARKER_BND_MISSING_WORKSPACE;
 import static org.bndtools.api.BndtoolsConstants.MARKER_BND_PATH_PROBLEM;
 import static org.bndtools.api.BndtoolsConstants.MARKER_BND_PROBLEM;
@@ -16,6 +18,7 @@ import java.util.Collection;
 import java.util.List;
 import java.util.Map.Entry;
 
+import org.bndtools.api.BndtoolsConstants;
 import org.bndtools.api.ILogger;
 import org.bndtools.api.IProjectValidator;
 import org.bndtools.api.IValidator;
@@ -27,6 +30,7 @@ import org.bndtools.build.api.MarkerData;
 import org.eclipse.core.resources.IMarker;
 import org.eclipse.core.resources.IProject;
 import org.eclipse.core.resources.IResource;
+import org.eclipse.core.resources.ResourcesPlugin;
 import org.eclipse.core.runtime.CoreException;
 import org.eclipse.core.runtime.IConfigurationElement;
 import org.eclipse.core.runtime.IStatus;
@@ -80,13 +84,35 @@ class MarkerSupport {
 
 	void deleteMarkers(String markerType) throws CoreException {
 		if (markerType.equals("*")) {
+			deleteMarkers(MARKER_BND_BLOCKER);
 			deleteMarkers(MARKER_BND_PROBLEM);
 			deleteMarkers(MARKER_BND_PATH_PROBLEM);
 			deleteMarkers(MARKER_BND_WORKSPACE_PROBLEM);
 			deleteMarkers(MARKER_BND_MISSING_WORKSPACE);
 			deleteMarkers(MARKER_JAVA_BASELINE);
-		} else
-			project.deleteMarkers(markerType, true, IResource.DEPTH_INFINITE);
+		} else {
+			if (project.exists() && project.isAccessible()) {
+				project.deleteMarkers(markerType, true, IResource.DEPTH_INFINITE);
+				deleteCnfMarkersForThisProject(markerType);
+			}
+		}
+	}
+
+	private void deleteCnfMarkersForThisProject(String markerType) throws CoreException {
+		IProject cnf = ResourcesPlugin.getWorkspace()
+			.getRoot()
+			.getProject("cnf");
+		if (cnf != null && cnf.isOpen()) {
+			IMarker[] markers = cnf.findMarkers(markerType, true, IResource.DEPTH_INFINITE);
+			if (markers != null) {
+				for (IMarker marker : markers) {
+					String projectName = marker.getAttribute(BndtoolsConstants.BNDTOOLS_MARKER_PROJECT_ATTR, null);
+					if (projectName == null || projectName.equals(project.getName())) {
+						marker.delete();
+					}
+				}
+			}
+		}
 	}
 
 	private void createMarkers(Processor model, int severity, Collection<String> msgs, String markerType)
@@ -104,32 +130,36 @@ class MarkerSupport {
 			BuildErrorDetailsHandler handler = BuildErrorDetailsHandlers.INSTANCE.findHandler(type);
 
 			List<MarkerData> markers = handler.generateMarkerData(project, model, location);
-			for (MarkerData markerData : markers) {
-				IResource resource = markerData.getResource();
-				if (resource != null && resource.exists()) {
-					String typeOverride = markerData.getTypeOverride();
-					IMarker marker = resource.createMarker(typeOverride != null ? typeOverride : markerType);
-					marker.setAttribute(IMarker.SEVERITY, severity);
-					marker.setAttribute("$bndType", type);
+			if (!markers.isEmpty()) {
+				for (MarkerData markerData : markers) {
+					IResource resource = markerData.getResource();
+					if (resource != null && resource.exists()) {
+						String typeOverride = markerData.getTypeOverride();
+						IMarker marker = resource.createMarker(typeOverride != null ? typeOverride : markerType);
+						marker.setAttribute(IMarker.SEVERITY, severity);
+						marker.setAttribute("$bndType", type);
 
-					//
-					// Set location information
-					if (location.header != null)
-						marker.setAttribute(BNDTOOLS_MARKER_HEADER_ATTR, location.header);
-					if (location.context != null)
-						marker.setAttribute(BNDTOOLS_MARKER_CONTEXT_ATTR, location.context);
-					if (location.file != null)
-						marker.setAttribute(BNDTOOLS_MARKER_FILE_ATTR, location.file);
-					if (location.reference != null)
-						marker.setAttribute(BNDTOOLS_MARKER_REFERENCE_ATTR, location.reference);
+						//
+						// Set location information
+						if (location.header != null)
+							marker.setAttribute(BNDTOOLS_MARKER_HEADER_ATTR, location.header);
+						if (location.context != null)
+							marker.setAttribute(BNDTOOLS_MARKER_CONTEXT_ATTR, location.context);
+						if (location.file != null)
+							marker.setAttribute(BNDTOOLS_MARKER_FILE_ATTR, location.file);
+						if (location.reference != null)
+							marker.setAttribute(BNDTOOLS_MARKER_REFERENCE_ATTR, location.reference);
 
-					marker.setAttribute(BuildErrorDetailsHandler.PROP_HAS_RESOLUTIONS, markerData.hasResolutions());
-					for (Entry<String, Object> attrib : markerData.getAttribs()
-						.entrySet())
-						marker.setAttribute(attrib.getKey(), attrib.getValue());
+						marker.setAttribute(BNDTOOLS_MARKER_PROJECT_ATTR, project.getName());
+
+						marker.setAttribute(BuildErrorDetailsHandler.PROP_HAS_RESOLUTIONS, markerData.hasResolutions());
+						for (Entry<String, Object> attrib : markerData.getAttribs()
+							.entrySet())
+							marker.setAttribute(attrib.getKey(), attrib.getValue());
+					}
 				}
+				return;
 			}
-			return;
 		}
 
 		String defaultResource = model instanceof Project ? Project.BNDFILE
@@ -157,6 +187,8 @@ class MarkerSupport {
 
 					if (dw.isTestBin(marker.getResource()))
 						continue;
+
+					System.out.println("error marker " + marker);
 					return true;
 				}
 			}
